+++ modules/sdk/gradle-plugins-workspace/src/main/java/com/liferay/gradle/plugins/workspace/WorkspaceExtension.java
@@ -20,6 +20,7 @@ import com.liferay.gradle.plugins.workspace.configurators.RootProjectConfigurato
 import com.liferay.gradle.plugins.workspace.configurators.ThemesProjectConfigurator;
 import com.liferay.gradle.plugins.workspace.configurators.WarsProjectConfigurator;
 import com.liferay.gradle.plugins.workspace.internal.util.GradleUtil;
+import com.liferay.gradle.util.Validator;
 import com.liferay.portal.tools.bundle.support.constants.BundleSupportConstants;
 
 import groovy.lang.MissingPropertyException;
@@ -49,6 +50,8 @@ public class WorkspaceExtension {
 		_projectConfigurators.add(new ThemesProjectConfigurator(settings));
 		_projectConfigurators.add(new WarsProjectConfigurator(settings));
 
+		_bundleCacheDir = _getProperty(
+			settings, "bundle.cache.dir", _BUNDLE_CACHE_DIR);
 		_bundleDistRootDirName = _getProperty(
 			settings, "bundle.dist.root.dir", _BUNDLE_DIST_ROOT_DIR_NAME);
 		_bundleTokenDownload = _getProperty(
@@ -60,6 +63,9 @@ public class WorkspaceExtension {
 			settings, "bundle.token.force", _BUNDLE_TOKEN_FORCE);
 		_bundleTokenPassword = _getProperty(
 			settings, "bundle.token.password", _BUNDLE_TOKEN_PASSWORD);
+		_bundleTokenPasswordFile = _getProperty(
+			settings, "bundle.token.password.file",
+			_BUNDLE_TOKEN_PASSWORD_FILE);
 		_bundleUrl = _getProperty(
 			settings, "bundle.url", BundleSupportConstants.DEFAULT_BUNDLE_URL);
 		_configsDir = _getProperty(
@@ -74,6 +80,10 @@ public class WorkspaceExtension {
 		_rootProjectConfigurator = new RootProjectConfigurator(settings);
 	}
 
+	public File getBundleCacheDir() {
+		return GradleUtil.toFile(_gradle.getRootProject(), _bundleCacheDir);
+	}
+
 	public String getBundleDistRootDirName() {
 		return GradleUtil.toString(_bundleDistRootDirName);
 	}
@@ -84,6 +94,11 @@ public class WorkspaceExtension {
 
 	public String getBundleTokenPassword() {
 		return GradleUtil.toString(_bundleTokenPassword);
+	}
+
+	public File getBundleTokenPasswordFile() {
+		return GradleUtil.toFile(
+			_gradle.getRootProject(), _bundleTokenPasswordFile);
 	}
 
 	public String getBundleUrl() {
@@ -128,6 +143,10 @@ public class WorkspaceExtension {
 		throw new MissingPropertyException(name, ProjectConfigurator.class);
 	}
 
+	public void setBundleCacheDir(Object bundleCacheDir) {
+		_bundleCacheDir = bundleCacheDir;
+	}
+
 	public void setBundleDistRootDirName(Object bundleDistRootDirName) {
 		_bundleDistRootDirName = bundleDistRootDirName;
 	}
@@ -146,6 +165,10 @@ public class WorkspaceExtension {
 
 	public void setBundleTokenPassword(Object bundleTokenPassword) {
 		_bundleTokenPassword = bundleTokenPassword;
+	}
+
+	public void setBundleTokenPasswordFile(Object bundleTokenPasswordFile) {
+		_bundleTokenPasswordFile = bundleTokenPasswordFile;
 	}
 
 	public void setBundleUrl(Object bundleUrl) {
@@ -171,12 +194,33 @@ public class WorkspaceExtension {
 			object, WorkspacePlugin.PROPERTY_PREFIX + keySuffix, defaultValue);
 	}
 
+	private Object _getProperty(
+		Object object, String keySuffix, File defaultValue) {
+
+		Object value = GradleUtil.getProperty(
+			object, WorkspacePlugin.PROPERTY_PREFIX + keySuffix);
+
+		if ((value instanceof String) && Validator.isNull((String)value)) {
+			value = null;
+		}
+
+		if (value == null) {
+			return defaultValue;
+		}
+
+		return value;
+	}
+
 	private String _getProperty(
 		Object object, String keySuffix, String defaultValue) {
 
 		return GradleUtil.getProperty(
 			object, WorkspacePlugin.PROPERTY_PREFIX + keySuffix, defaultValue);
 	}
+
+	private static final File _BUNDLE_CACHE_DIR = new File(
+		System.getProperty("user.home"),
+		BundleSupportConstants.DEFAULT_BUNDLE_CACHE_DIR_NAME);
 
 	private static final String _BUNDLE_DIST_ROOT_DIR_NAME = null;
 
@@ -188,11 +232,15 @@ public class WorkspaceExtension {
 
 	private static final String _BUNDLE_TOKEN_PASSWORD = null;
 
+	private static final String _BUNDLE_TOKEN_PASSWORD_FILE = null;
+
+	private Object _bundleCacheDir;
 	private Object _bundleDistRootDirName;
 	private Object _bundleTokenDownload;
 	private Object _bundleTokenEmailAddress;
 	private Object _bundleTokenForce;
 	private Object _bundleTokenPassword;
+	private Object _bundleTokenPasswordFile;
 	private Object _bundleUrl;
 	private Object _configsDir;
 	private Object _environment;