+++ modules/sdk/gradle-plugins-workspace/src/main/java/com/liferay/gradle/plugins/workspace/configurators/RootProjectConfigurator.java
@@ -238,6 +238,16 @@ public class RootProjectConfigurator implements Plugin<Project> {
 
 			});
 
+		createTokenTask.setPasswordFile(
+			new Callable<File>() {
+
+				@Override
+				public File call() throws Exception {
+					return workspaceExtension.getBundleTokenPasswordFile();
+				}
+
+			});
+
 		return createTokenTask;
 	}
 
@@ -317,13 +327,6 @@ public class RootProjectConfigurator implements Plugin<Project> {
 		final Download download = GradleUtil.addTask(
 			project, DOWNLOAD_BUNDLE_TASK_NAME, Download.class);
 
-		File destinationDir = new File(
-			System.getProperty("user.home"), ".liferay/bundles");
-
-		destinationDir.mkdirs();
-
-		download.dest(destinationDir);
-
 		download.doFirst(
 			new Action<Task>() {
 
@@ -353,6 +356,13 @@ public class RootProjectConfigurator implements Plugin<Project> {
 					if (workspaceExtension.isBundleTokenDownload()) {
 						download.dependsOn(createTokenTask);
 					}
+
+					File destinationDir =
+						workspaceExtension.getBundleCacheDir();
+
+					destinationDir.mkdirs();
+
+					download.dest(destinationDir);
 
 					Object src = download.getSrc();