+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checkstyle/checks/MissingOverrideCheck.java
@@ -16,8 +16,10 @@ package com.liferay.source.formatter.checkstyle.checks;
 
 import com.liferay.portal.kernel.util.CharPool;
 import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.util.SetUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Tuple;
+import com.liferay.source.formatter.SourceFormatterExcludes;
 import com.liferay.source.formatter.checks.util.SourceUtil;
 import com.liferay.source.formatter.util.SourceFormatterUtil;
 import com.liferay.source.formatter.util.ThreadSafeSortedClassLibraryBuilder;
@@ -168,7 +170,8 @@ public class MissingOverrideCheck extends AbstractCheck {
 		}
 
 		List<String> fileNames = SourceFormatterUtil.scanForFiles(
-			absolutePath + "/", _EXCLUDES, new String[] {"**/*.java"}, true);
+			absolutePath + "/", new String[0], new String[] {"**/*.java"},
+			new SourceFormatterExcludes(SetUtil.fromArray(_EXCLUDES)), true);
 
 		for (String curFileName : fileNames) {
 			curFileName = StringUtil.replace(
@@ -348,7 +351,7 @@ public class MissingOverrideCheck extends AbstractCheck {
 		return false;
 	}
 
-	private static final String[] _EXCLUDES = new String[] {
+	private static final String[] _EXCLUDES = {
 		"**/.git/**", "**/.gradle/**", "**/bin/**", "**/build/**",
 		"**/classes/**", "**/node_modules/**", "**/npm-shrinkwrap.json",
 		"**/package-lock.json", "**/test-classes/**", "**/test-coverage/**",