+++ modules/sdk/gradle-plugins-workspace/CHANGELOG.markdown
@@ -128,6 +128,25 @@ version 3.0.2.
 - [LPS-73746]: Trim authentication token in case users add extra lines into
 their token file.
 
+## 1.6.0 - 2017-08-01
+
+### Added
+- [LPS-73913]: Add the ability to pass the Liferay bundle authentication token
+password from a file by setting the
+`liferay.workspace.bundle.token.password.file` property in `gradle.properties`.
+
+### Changed
+- [LPS-73855]: Update the [Liferay Gradle Plugins] dependency to version 3.4.17.
+- [LPS-73913]: Update the [Liferay Portal Tools Bundle Support] dependency to
+version 3.1.0.
+
+## 1.7.0 - *(Unreleased)*
+
+### Added
+- [LPS-74124]: Add the ability to configure the cache directory for downloaded
+Liferay bundles by setting the `liferay.workspace.bundle.cache.dir` property in
+`gradle.properties`.
+
 [Gradle Download Task]: https://github.com/michel-kraemer/gradle-download-task/
 [Liferay Gradle Plugins]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins
 [Liferay Portal Tools Bundle Support]: https://github.com/liferay/liferay-portal/tree/master/modules/util/portal-tools-bundle-support
@@ -156,3 +175,6 @@ their token file.
 [LPS-73248]: https://issues.liferay.com/browse/LPS-73248
 [LPS-73383]: https://issues.liferay.com/browse/LPS-73383
 [LPS-73746]: https://issues.liferay.com/browse/LPS-73746
+[LPS-73855]: https://issues.liferay.com/browse/LPS-73855
+[LPS-73913]: https://issues.liferay.com/browse/LPS-73913
+[LPS-74124]: https://issues.liferay.com/browse/LPS-74124